refactor: modularize charge sensor compensation routines and introduce RANSAC fitting utilities
- Moved RANSAC fitting logic into a new `ransac_fitting` module for better organization and reusability.
- Removed the RANSACFitResult class from the charge sensor compensation module and integrated it into the new module.
- Updated the `charge_sensor_compensation` module to utilize the new RANSAC fitting utilities, streamlining the compensation gradient calculation process.
- Enhanced documentation for the new module and updated references in existing routines to reflect the changes.
- Improved overall structure of charge sensor routines for better maintainability and clarity.
