+++ tests/routines/builtins/test_dqd_search.py
